Update Hopper CalibrateLift command, attempt to fix
Updated the Hopper CalibrateLift command to work with the updated enum from the previous commit, and attempted to make it work correctly by changing some values.

See also: #27
